random commit, join my dc https://discord.gg/ZuUWPaSrHa

+++ addons/GodotSTG/resources/STGBulletModifier.cs
@@ -13,7 +13,7 @@ public partial class STGBulletModifier:Resource{
     [Export] public Type type {get; set;} = Type.Regular;
     [Export] public float speed {get; set;}
     [Export] public Vector2 acceleration {get; set;}
-    [Export] public float lifespan {get; set;} = -1;
+    [Export] public double lifespan {get; set;} = -1;
     [Export] public float sine_freq {get; set;}
     [Export] public float sine_width {get; set;}
     [Export] public STGBulletModifier zoned {get; set;}
